# swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/pancake/connect": {
      "post": {
        "operationId": "PancakeController_connect",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/status": {
      "get": {
        "operationId": "PancakeController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/disconnect": {
      "delete": {
        "operationId": "PancakeController_disconnect",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages": {
      "get": {
        "operationId": "PancakeController_listPages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages/{pageId}/leads": {
      "get": {
        "operationId": "PancakeController_listLeads",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages/{pageId}/sync": {
      "post": {
        "operationId": "PancakeController_syncPage",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages/{pageId}/auto-label": {
      "post": {
        "operationId": "PancakeController_autoLabel",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/leads/{leadId}": {
      "patch": {
        "operationId": "PancakeController_updateLead",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/leads/{leadId}/mark-customer": {
      "post": {
        "operationId": "PancakeController_markCustomer",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages/{pageId}/conversations": {
      "get": {
        "operationId": "PancakeController_listConversations",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages/{pageId}/conversations/{conversationId}/messages": {
      "get": {
        "operationId": "PancakeController_listMessages",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/pages/{pageId}/conversations/{conversationId}/lead-preview": {
      "get": {
        "operationId": "PancakeController_leadPreview",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "pancake"
        ]
      }
    },
    "/api/v1/pancake/webhook": {
      "post": {
        "operationId": "PancakeWebhookController_webhook",
        "parameters": [
          {
            "name": "x-pancake-secret",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "secret",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "pancake-webhook"
        ]
      },
      "get": {
        "operationId": "PancakeWebhookController_webhookPing",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pancake-webhook"
        ]
      }
    }
  },
  "info": {
    "title": "CQA CRM Backend API",
    "description": "Tài liệu API dành cho hệ thống CQA CRM (Customer Quality Audit)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Nhập JWT Token để truy cập các API cần xác thực",
        "in": "header"
      }
    },
...
  }
}